+++ pyninja/database.py
@@ -1,0 +1,48 @@
+from pyninja import models
+
+
+def get_record(host: str) -> int | None:
+    """Gets blocked epoch time for a particular host.
+
+    Args:
+        host: Host address.
+
+    Returns:
+        int:
+        Returns the epoch time until when the host address should be blocked.
+    """
+    with models.database.connection:
+        cursor = models.database.connection.cursor()
+        state = cursor.execute(
+            "SELECT block_until FROM auth_errors WHERE host=(?)", (host,)
+        ).fetchone()
+    if state and state[0]:
+        return state[0]
+
+
+def put_record(host: str, block_until: int) -> None:
+    """Inserts blocked epoch time for a particular host.
+
+    Args:
+        host: Host address.
+        block_until: Epoch time until when the host address should be blocked.
+    """
+    with models.database.connection:
+        cursor = models.database.connection.cursor()
+        cursor.execute(
+            "INSERT INTO auth_errors (host, block_until) VALUES (?,?)",
+            (host, block_until),
+        )
+        models.database.connection.commit()
+
+
+def remove_record(host: str) -> None:
+    """Deletes all records related to the host address.
+
+    Args:
+        host: Host address.
+    """
+    with models.database.connection:
+        cursor = models.database.connection.cursor()
+        cursor.execute("DELETE FROM auth_errors WHERE host=(?)", (host,))
+        models.database.connection.commit()