+++ pyninja/__init__.py
@@ -7,7 +7,7 @@ import click
 
 from pyninja.main import start  # noqa: F401
 
-version = "0.0.0"
+version = "0.0.1"
 
 
 @click.command()