+++ release_notes.rst
@@ -3,7 +3,7 @@ Release Notes
 
 0.0.0 (08/10/2024)
 ------------------
-- Release first stable version
+- Release stable version
 
 0.0.0-a (08/10/2024)
 --------------------