+++ release_notes.rst
@@ -1,9 +1,18 @@
 Release Notes
 =============
 
-0.0.0 (08/10/2024)
-------------------
-- Release stable version
+v0.0.1 (08/11/2024)
+-------------------
+- Includes a process monitor and remote command execution functionality
+- Security improvements including brute force protection and rate limiting
+- Accepts ``JSON`` and ``YAML`` files for env config
+- Supports custom worker count for ``uvicorn`` server
+- Allows custom logging using ``logging.ini``
+- Includes an option to set the ``apikey`` via commandline
+
+v0.0.0 (08/11/2024)
+-------------------
+- Release first stable version
 
 0.0.0-a (08/10/2024)
 --------------------